Cambios menores para acabar la primera parte del ejercicio antes de implementar retrofit
