+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,13 @@
             android:name=".MainActivity"
             android:exported="true" />
         <activity
-            android:name="ut2.UiActivity"
+            android:name="com.tijanidian.mvvm.ut2.UiActivity"
             android:exported="true" />
         <activity
-            android:name="ut2.UiTestGlovoActivity"
+            android:name="com.tijanidian.mvvm.ut2.UiTestGlovoActivity"
             android:exported="true"/>
             <activity
-                android:name="com.tijanidian.mvvm.presentation.AlertActivity"
+                android:name="com.tijanidian.alerts.presentation.AlertActivity"
                 android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
