+++ app/src/main/java/com/tijanidian/alerts/domain/AlertRepository.kt
@@ -1,0 +1,5 @@
+package com.tijanidian.alerts.domain
+
+interface AlertRepository {
+    fun getAlerts(): List<AlertModel>
+}
