+++ app/src/main/java/com/tijanidian/alerts/domain/GetAlertUseCase.kt
@@ -1,0 +1,8 @@
+package com.tijanidian.alerts.domain
+
+class GetAlertUseCase(private val alertRepository: AlertRepository) {
+
+    fun execute(): List<AlertModel> {
+        return alertRepository.getAlerts()
+    }
+}
