--- app/src/main/java/com/tijanidian/mvvm/domain/AlertRepository.kt
@@ -1,5 +1,0 @@
-package com.tijanidian.mvvm.domain
-
-interface AlertRepository {
-    fun getAlerts() : List<AlertModel>
-}
