--- app/src/main/java/com/tijanidian/mvvm/domain/GetAlertUseCase.kt
@@ -1,8 +1,0 @@
-package com.tijanidian.mvvm.domain
-
-class GetAlertUseCase(val alertRepository: AlertRepository){
-
-    fun execute(): List<AlertModel> {
-        return alertRepository.getAlerts()
-    }
-}
