+++ app/src/main/java/com/tijanidian/mvvm/ut2/UiTestGlovoActivity.kt
@@ -1,4 +1,4 @@
-package ut2
+package com.tijanidian.mvvm.ut2
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
