+++ app/src/main/java/com/tijanidian/mvvm/ut2/UiTestGlovoViewModel.kt
@@ -1,4 +1,4 @@
-package ut2
+package com.tijanidian.mvvm.ut2
 
 class UiTestGlovoViewModel {
     fun getGlovoModel() = GlovoModel()
