+++ app/src/main/res/layout/activity_alerts.xml
@@ -1,50 +1,58 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:card_view="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/card_view"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginTop="100dp"
-    tools:context="com.tijanidian.mvvm.presentation.AlertActivity">
+    android:layout_height="wrap_content"
+    android:layout_margin="@dimen/margin_left"
+    card_view:cardCornerRadius="@dimen/corner_card">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context="com.tijanidian.alerts.presentation.AlertActivity">
+
+        <ImageView
+            android:id="@+id/icon_warning"
+            android:layout_width="@dimen/icon_width_medium"
+            android:layout_height="@dimen/icon_height_medium"
+            app:layout_constraintEnd_toStartOf="@+id/main_title"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            tools:src="@drawable/ic_warning" />
+
+        <TextView
+            android:id="@+id/main_title"
+            style="@style/main_text_plag"
+            android:layout_width="@dimen/max_height_plag"
+            android:layout_height="wrap_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            tools:text="@string/text_main_plag" />
+
+        <TextView
+            android:id="@+id/date_published"
+            android:layout_width="@dimen/max_height_plag"
+            android:layout_height="wrap_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/main_title"
+            tools:text="@string/text_date_plag" />
+
+        <TextView
+            android:id="@+id/summary"
+            style="@style/text_plag"
+            android:layout_width="@dimen/max_height_plag"
+            android:layout_height="wrap_content"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/date_published"
+            tools:text="@string/text_body_plag" />
 
 
-    <ImageView
-        android:id="@+id/image_view"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        tools:src="@drawable/ic_warning"
-        app:layout_constraintEnd_toStartOf="@+id/text_bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</androidx.cardview.widget.CardView>
 
-    <TextView
-        android:id="@+id/text_bold"
-        android:layout_width="320dp"
-        android:layout_height="wrap_content"
-        style="@style/main_text_plag"
-        tools:text="@string/text_bold_plag"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/current_date"
-        android:layout_width="320dp"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/text_bold"
-        tools:text="@string/text_date_plag" />
-
-    <TextView
-        android:id="@+id/text_body"
-        android:layout_width="320dp"
-        style="@style/text_plag"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/current_date"
-        tools:text="@string/text_body_plag" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
